next

+++ build-dpdk.sh
@@ -8,9 +8,15 @@ set -eux
 
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 
-yum install xz meson -y || true
-yum install "@Development Tools" -y || true
-yum install python3-pyelftools.noarch -y || true
+# yum install xz meson -y || true
+# yum install "@Development Tools" -y || true
+# yum install python3-pyelftools.noarch -y || true
+
+if [ -f "${SCRIPT_DIR}/build-dir/dpdk-ethtool" ] && [ -f "${SCRIPT_DIR}/build-dir/dpdk-testpmd" ]; then
+  echo "dpdk-ethtool and dpdk-testpmd binaries already exist."
+  echo "Run 'make clean' if you wish to force a rebuild."
+  exit 0
+fi
 
 rm -Rf "${SCRIPT_DIR}/build-dir"
 mkdir -p "${SCRIPT_DIR}/build-dir"
@@ -26,4 +32,5 @@ meson configure --buildtype=debug --optimization 0 -Dexamples=all
 ninja
 cd "${SCRIPT_DIR}/build-dir"
 cp dpdk/build/app/dpdk-testpmd .
+cp ./dpdk/build/examples/dpdk-ethtool .
 rm -Rf dpdk
